Fix:move Commented_List.js and Commented_List.css to Community directory and fix related path in some files. #104 #91 #93
